Move new_line_formatter.py to formatter.py for future formatters
