+++ vllm/logging/__init__.py
@@ -1,4 +1,4 @@
-from vllm.logging.new_line_formatter import NewLineFormatter
+from vllm.logging.formatter import NewLineFormatter
 
 __all__ = [
     "NewLineFormatter",
